return original useArcgis
